requirements.txt
flask
face_recognition
opencv-python-headless
numpy
pillow